Send the content-md5 header for multipart uploads
When uploading each part of a multipart upload, we can send the
`content-md5` header with the md5 digest of the chunk. When uploading,
if the chunks contents do not match the md5 we have provided we'll be
returned an error from AWS which will allow us to retry uploading the
chunk.